重要升级
更新（2015.03.24）：
1、AppConfig.h 文件添加: #define AiitecOpenAPI_DEBUG 0
2、Supporting Classes 去掉中间的空格，变为 SupportingClasses
3、SupportingClasses/AIIPacketRequest/PacketRequest.h 添加类方法
+ (void)packetConnectionFinishedWarningHandle:(AIIPacketConnection
*)connection;
用于处理:
#pragma mark - PacketHttpConnectionDelegate
- (void)packetConnectionFinished:(AIIPacketConnection *)connection｛
    if (connection.response.query.status) {
        [PacketRequest
packetConnectionFinishedWarningHandle:connection];
    ｝
｝
4、修复底层各种警告

+++ AiitecOpenAPISource/AiitecOpenAPISource/AiitecOpenAPI/AIIPacketRequest/AIIPacketRequest.m
@@ -58,13 +58,13 @@
             [AIIPacketRequest packetSetting:AIIPacketSettingViaWWAN value:[NSNumber numberWithBool:YES]];
         }
     });
-    return packetRequest.settingDictionary[[NSString stringWithFormat:@"%lu", key]];
+    return packetRequest.settingDictionary[[NSString stringWithFormat:@"%lu", (unsigned long)key]];
 }
 
 + (void)packetSetting:(AIIPacketSetting)key value:(id)value;
 {
     AIIPacketRequest *packetRequest = [AIIPacketRequest sharedAIIPacketRequest];
-    packetRequest.settingDictionary[[NSString stringWithFormat:@"%lu", key]] = value;
+    packetRequest.settingDictionary[[NSString stringWithFormat:@"%lu", (unsigned long)key]] = value;
     [packetRequest.settingDictionary writeToFile:[AIIPacketRequest settingPlistPath] atomically:YES];
 }
